Move `ExtensionGeneratorTest` and extract path cleanup logic to trait.
